Remove redundant type casts on size_of and function pointer Options
- Remove `as usize` from `core::mem::size_of::<T>()` (already returns usize)
- Remove `as Option<FnType>` outer casts from `Some(fn as FnType)`
- Remove `as lzma_*_function` type alias casts from `Some(fn as FnType)`

+++ liblzma-rs/src/common/filter_decoder.rs
@@ -366,8 +366,8 @@ static decoders: [lzma_filter_decoder; 12] = [
 extern "C" fn decoder_find(id: lzma_vli) -> *const lzma_filter_decoder {
     let mut i: size_t = 0;
     while i
-        < (core::mem::size_of::<[lzma_filter_decoder; 12]>() as usize)
-            .wrapping_div(core::mem::size_of::<lzma_filter_decoder>() as usize)
+        < (core::mem::size_of::<[lzma_filter_decoder; 12]>())
+            .wrapping_div(core::mem::size_of::<lzma_filter_decoder>())
     {
         if decoders[i as usize].id == id {
             return decoders.as_ptr().wrapping_add(i as usize);
